Renaming initialize folder
Python gets confused if the subfolder is the same as a file name in the same folder, so the initialize folder is being renamed to setup
