+++ dataactcore/scripts/setup/load_object_class.py
