+++ dataactcore/scripts/setup/read_zips.py
