+++ dataactcore/scripts/setup/setup_static_data.py
