+++ dataactcore/scripts/setup/setup_submission_type_db.py
