+++ tests/integration/baseTestAPI.py
@@ -9,15 +9,15 @@ from dataactcore.interfaces.db import GlobalDB
 from dataactcore.interfaces.function_bag import create_user_with_password, get_password_hash
 from dataactcore.models.domainModels import CGAC, FREC, SubTierAgency
 from dataactcore.models.userModel import User, UserAffiliation
-from dataactcore.scripts.initialize.database_setup import drop_database
-from dataactcore.scripts.initialize.setup_user_db import setup_user_db
-from dataactcore.scripts.initialize.setup_job_tracker_db import setup_job_tracker_db
-from dataactcore.scripts.initialize.setup_error_db import setup_error_db
-from dataactcore.scripts.initialize.setup_validation_db import setup_validation_db
-from dataactcore.scripts.initialize.database_setup import create_database, run_migrations
+from dataactcore.scripts.setup.database_setup import drop_database
+from dataactcore.scripts.setup.setup_user_db import setup_user_db
+from dataactcore.scripts.setup.setup_job_tracker_db import setup_job_tracker_db
+from dataactcore.scripts.setup.setup_error_db import setup_error_db
+from dataactcore.scripts.setup.setup_validation_db import setup_validation_db
+from dataactcore.scripts.setup.database_setup import create_database, run_migrations
 from dataactcore.config import CONFIG_BROKER, CONFIG_DB
 import dataactcore.config
-from dataactcore.scripts.initialize.setup_emails import setup_emails
+from dataactcore.scripts.setup.setup_emails import setup_emails
 from dataactvalidator.health_check import create_app as create_validator_app
 from dataactcore.models.lookups import ALL_PERMISSION_TYPES_DICT
 from tests.unit.dataactcore.factories.user import UserFactory
